Login com persistência + atualização rota
