+++ src/context/AuthProvider.jsx
@@ -3,21 +3,21 @@ import { login as loginService } from "@/services/authAPI"
 import { useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
 import { getUserData } from "@/services/adotanteAPI";
+import { axiosAuth } from "@/services/configs/axiosConfig";
 
 const AuthContext = createContext();
 
 export const AuthProvider = ({ children }) => {
     const [auth, setAuth] = useState({
-        idUser: null,
         token: localStorage.getItem("token") || null,
-        userData: JSON.parse(localStorage.getItem("userData")) || null, // Tem que mudar
+        userData: null,
     });
 
     const navigate = useNavigate();
 
-    const retriveUserData = async (id) => {
+    const retriveUserData = async () => {
         try {
-            const response = await getUserData(id);
+            const response = await getUserData();
             setAuth((prevAuth) => ({
                 ...prevAuth,
                 userData: response.data
@@ -31,11 +31,10 @@ export const AuthProvider = ({ children }) => {
     const login = async (context, data) => {
         try {
             const response = await loginService(context, data);
-            const { token, idUser } = response.data;
+            const { token } = response.data;
 
             setAuth((prevAuth) => ({
                 ...prevAuth,
-                idUser,
                 token,
             }));
 
@@ -43,7 +42,7 @@ export const AuthProvider = ({ children }) => {
 
             localStorage.setItem("token", token);
 
-            await retriveUserData(idUser);
+            await retriveUserData();
 
             navigate("/");
         } catch (error) {
@@ -53,20 +52,67 @@ export const AuthProvider = ({ children }) => {
     }
 
     const logout = () => {
-        setAuth({ idUser: null, token: null });
+        setAuth({ token: null });
         localStorage.removeItem("token");
-        localStorage.removeItem("userData");
         navigate("/login");
     };
 
+    const refreshToken = async () => {
+        try {
+            const refreshToken = localStorage.getItem("token");
+            const response = await axiosAuth.post("/login/refresh", {
+                refreshToken,
+            });
+
+            const { token } = response.data;
+            localStorage.setItem("token", token);
+
+            setAuth((prevAuth) => ({
+                ...prevAuth,
+                token,
+            }));
+
+            return token;
+        } catch (refreshError) {
+            console.error("Erro ao tentar renovar o token", refreshError);
+            logout();
+            throw refreshError;
+        }
+    };
+
+    axiosAuth.interceptors.response.use(
+        (response) => response,
+        async (error) => {
+            const originalRequest = error.config;
+
+            if (error.response.status === 401 && !originalRequest._retry) {
+                originalRequest._retry = true;
+
+                try {
+                    const newToken = await refreshToken();
+                    originalRequest.headers.Authorization = `Bearer ${newToken}`;
+                    return axiosAuth(originalRequest);
+                } catch (refreshError) {
+                    return Promise.reject(refreshError);
+                }
+            }
+
+            return Promise.reject(error);
+        }
+    );
+
+
     useEffect(() => {
         const token = localStorage.getItem("token");
-        if (token && !auth.token || auth.userData) {
+        if (token && !auth.token) {
             setAuth((prev) => ({ ...prev, token }));
-            localStorage.setItem("userData", JSON.stringify(auth.userData)); // Estudando outra solução
         }
-    }, [auth.token, auth.userData]);
 
+        if (auth.token && !auth.userData) {
+            retriveUserData();
+        }
+
+    }, [auth.token, auth.userData]);
 
     return (
         <AuthContext.Provider value={{ auth, setAuth, login, logout }}>
